backend/requirements.txt
# Backend dependencies for Personal AI Reading Companion

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.2.0
openai==1.51.0
gTTS==2.5.0
aiofiles==23.2.1
pydantic==2.6.0
httpx==0.27.2
